Replace endpoint for generic relayers
