+++ .env.production
@@ -1,4 +1,4 @@
-WORMSCAN_API_BASE_URL="https://api.wormscan.io/api/v1"
+WORMSCAN_API_BASE_URL="https://api.wormholescan.io/api/v1"
 WORMSCAN_TESTNET_API_BASE_URL="https://api.testnet.wormholescan.io/api/v1"
 WORMHOLE_MARKET_TOKENS_URL="https://raw.githubusercontent.com/certusone/wormhole-token-list/main/src/markets.json"
